Move supporting objects.
