+++ lib/active_triples/configuration.rb
@@ -1,4 +1,7 @@
 module ActiveTriples
+  require_relative 'configuration/reflection'
+  require_relative 'configuration/merge_reflection'
+  require_relative 'configuration/reflection_factory'
   class Configuration
     attr_accessor :inner_hash
     def initialize(options={})
@@ -7,17 +10,16 @@ module ActiveTriples
 
     def merge(options)
       new_config = Configuration.new(options)
-      new_config.properties.each do |property, reflection|
-        current_property = properties[property] || reflection_factory.new(self, property)
-        current_property.set reflection.value
+      new_config.reflections.each do |property, reflection|
+        self.build_reflection(property).set reflection.value
       end
       self
     end
     
-    def properties
+    def reflections
       to_h.each_with_object({}) do |config_value, hsh|
         key = config_value.first
-        hsh[key] = reflection_factory.new(self, key)
+        hsh[key] = build_reflection(key)
       end
     end
 
@@ -29,6 +31,12 @@ module ActiveTriples
       @inner_hash.slice(*valid_config_options)
     end
 
+    protected
+
+    def build_reflection(key)
+      reflection_factory.new(self, key)
+    end
+
     private
 
     def reflection_factory
@@ -37,51 +45,6 @@ module ActiveTriples
 
     def valid_config_options
       [:base_uri, :rdf_label, :type, :repository]
-    end
-
-    class Reflection
-      attr_reader :object, :key
-      def initialize(object, key)
-        @object = object
-        @key = key
-      end
-
-      def value
-        object.inner_hash[key]
-      end
-
-      def set(value)
-        object.inner_hash[key] = value
-      end
-    end
-
-    class MergeReflection < Reflection
-      def set(value)
-        object.inner_hash[key] = Array(object.inner_hash[key])
-        object.inner_hash[key] |= Array(value)
-      end
-    end
-
-    class ReflectionFactory
-      def new(object, name)
-        if merge_configs.include?(name)
-          merge_reflection.new(object, name)
-        else
-          reflection.new(object, name)
-        end
-      end
-
-      def merge_reflection
-        MergeReflection
-      end
-
-      def reflection
-        Reflection
-      end
-
-      def merge_configs
-        [:type]
-      end
     end
   end
