+++ lib/active_triples/configuration/merge_reflection.rb
@@ -1,0 +1,10 @@
+module ActiveTriples
+  class Configuration
+    class MergeReflection < Reflection
+      def set(value)
+        object.inner_hash[key] = Array(object.inner_hash[key])
+        object.inner_hash[key] |= Array(value)
+      end
+    end
+  end
+end
