+++ lib/active_triples/configuration/reflection.rb
@@ -1,0 +1,19 @@
+module ActiveTriples
+  class Configuration
+    class Reflection
+      attr_reader :object, :key
+      def initialize(object, key)
+        @object = object
+        @key = key
+      end
+
+      def value
+        object.inner_hash[key]
+      end
+
+      def set(value)
+        object.inner_hash[key] = value
+      end
+    end
+  end
+end
